+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### BUG FIXES:
 
-- fix(block_fastly_service_product_enablement): remove the vcl check on product enablement for bot management ([#1270](https://github.com/fastly/terraform-provider-fastly/pull/1270))
+- fix(block_fastly_service_product_enablement): Allow `bot_management` to be enabled on Compute services ([#1270](https://github.com/fastly/terraform-provider-fastly/pull/1270))
 
 ### DEPENDENCIES:
 - build(deps): `github.com/hashicorp/terraform-plugin-sdk/v2` from 2.40.0 to 2.40.1 ([#1259](https://github.com/fastly/terraform-provider-fastly/pull/1259))