+++ CHANGELOG.md
@@ -4,6 +4,16 @@
 
 ### ENHANCEMENTS:
 
+### BUG FIXES:
+
+### DEPENDENCIES:
+
+### DOCUMENTATION:
+
+## 9.2.0 (May 20, 2026)
+
+### ENHANCEMENTS:
+
 - feat(service/backend): added support for the `MaxLifetime` and `MaxUse` attributes ([#1233](https://github.com/fastly/terraform-provider-fastly/pull/1233))
 
 ### BUG FIXES:
@@ -16,8 +26,6 @@
 - build(deps): `github.com/hashicorp/terraform-plugin-sdk/v2` from 2.40.0 to 2.40.1 ([#1259](https://github.com/fastly/terraform-provider-fastly/pull/1259))
 - build(deps): `github.com/fastly/go-fastly/v15` from 14.0.2 to 15.0.1([#1260](https://github.com/fastly/terraform-provider-fastly/pull/1260))
 - build(deps): `golang.org/x/net` from 0.53.0 to 0.54.0 ([#1267](https://github.com/fastly/terraform-provider-fastly/pull/1267))
-
-### DOCUMENTATION:
 
 ## 9.1.1 (April 22, 2026)